Extrait la définition d'une enum
pour éviter une erreur de compilation lors de sa réutilisation
